# openapi.yaml
openapi: 3.0.3
info:
  title: Mercado Uno API
  version: 0.2.0
  description: |
    API REST del marketplace educativo Mercado Uno. PostgreSQL es la base de
    producción. El checkout reserva inventario y crea un pedido y un pago
    pendientes; el proveedor simulado solo está disponible fuera de producción.
  license:
    name: Uso educativo
security: []
servers:
  - url: http://localhost:3000
    description: Stack Docker, con la API publicada por el proxy del frontend
  - url: http://localhost:4000
    description: API directa para desarrollo
...
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: Listar productos
      operationId: listProducts
      responses:
        "200":
          description: Catálogo completo.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/products/{id}:
    get:
      tags: [Products]
      summary: Obtener un producto
      operationId: getProductById
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Producto encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/InvalidIdentifier"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/cart:
    get:
      tags: [Cart]
      summary: Consultar el carrito propio
      operationId: getCart
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Líneas del carrito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItemDetailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Cart]
      summary: Añadir una cantidad al carrito
      operationId: addCartItem
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartRequest"
      responses:
        "200":
          description: Carrito actualizado.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItemSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InvalidOrigin"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Cart]
      summary: Reemplazar la cantidad de una línea
      operationId: updateCartItem
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartRequest"
      responses:
        "200":
          description: Cantidad actualizada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InvalidOrigin"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/cart/{productId}:
    delete:
      tags: [Cart]
      summary: Eliminar una línea del carrito
      operationId: deleteCartItem
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CartProductId"
      responses:
        "200":
          description: Línea eliminada o ya ausente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/InvalidIdentifier"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InvalidOrigin"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mercado_session
      description: En producción se usa el nombre endurecido __Host-mercado_session.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Compatibilidad temporal para clientes no navegador.
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
    CartProductId:
      name: productId
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
  responses:
    BadRequest:
      description: Entrada inválida.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidIdentifier:
      description: El identificador no es un entero positivo.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Sesión ausente, inválida o expirada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidOrigin:
      description: Una mutación con cookie llegó desde un origen no autorizado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: El recurso no existe o no pertenece al usuario.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    TooManyRequests:
      description: Se superó el límite temporal de solicitudes.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    SuccessResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, enum: [true] }
...
    Product:
      type: object
      required: [id, title, price, stock]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string, nullable: true }
        price: { type: number, format: double, example: 9.99 }
        stock: { type: integer, minimum: 0 }
    CartRequest:
      type: object
      required: [productId, qty]
      properties:
        productId: { type: integer, minimum: 1 }
        qty: { type: integer, minimum: 1, maximum: 100 }
    CartItemSummary:
      type: object
      required: [productId, qty]
      properties:
        productId: { type: integer }
        qty: { type: integer }
    CartItemDetailed:
      allOf:
        - $ref: "#/components/schemas/CartItemSummary"
        - type: object
          required: [product]
          properties:
            product:
              $ref: "#/components/schemas/Product"
...